refactor: InsatnceSet refactor
